pyproject.toml
[project]
name = "pdf2md"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai>=0.8.3",
    "pdf2image>=1.17.0",
    "pillow>=11.0.0",
    "pypeln>=0.4.9",
    "requests>=2.32.3",
    "setuptools>=75.6.0",
    "tqdm>=4.67.1",
]
